Move live updates logic to hook
